rename image related tests
