+++ tests/test_image_imarith.py
