+++ tests/test_image_imcombine.py
